Fix race between maybeReserve and memory abort (#5902)
Summary:
There is a race condition between maybeReserve and memory abort which
leads the segment fault during the operator execution. The following is the
event sequence of this race condition:
T1: aggregation operator request memory arbitration by maybeReserve
T2: memory arbitration fails for whatever reason that cause this memory
       pool to be aborted.
T3: the memory arbitration aborts the query and wait for the query execution
       to stop, that is, no running threads. Note the aggregation driver thread
       which initiates the arbitration is under driver suspension state so it is not
       counted as a running thread.
T4: the memory arbitration abort free up the memory resource from all the
       operators of the aborted query as it has been totally stopped.
T5: however the aggregation thread is not stopped and after the memory
      arbitration fails, maybeReserve will catch the exception and returns false
T6: the aggregation operator will continue the execution after maybeReserve
      fails by triggering disk spilling.
T7: the aggregation operator runs into random segment fault.

We shall throw the exception in maybeReserve if the query pool has been aborted
to stop the query execution. This PR reproduce this race condition and verified
the fix.
In followup, once the integration of memory arbitration and disk spilling becomes
stable. We shall remove the disk spilling after maybeReserve fails as the memory
arbitration is always able to spill if needs. We might want to keep the threshold
triggered disk spilling as it might be useful to handle some special query shape as
we have found in Meta internal staging tests.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5902

Reviewed By: mbasmanova

Differential Revision: D47860804

Pulled By: xiaoxmeng

fbshipit-source-id: 7f098e1e8da6d40bfeedb1281d843bd79d246900