Round large non-contiguous HashStringAllocator allocs to standarde sizes (#5921)
Summary:
When a non-contiguous allocation is requested and there are no free blocks in an HSA, we append a new slab of the requested sizes. This slab is rounded up to a standard size in the AllocationPool. We use the requested size instead of the rounded up size to make the new slab. This wastes space at the tail, e.g. we ask for 200K and get 256K. Furthermore when the HSA adds new slabs after this, it fails checkConsistency() because checkConsistency() reads the ranges from allocationPool and expects them to be entirely filled by HSA arenas, This is not the case here because the arena that was added for the 200K string is 200K and not the range of 256K.

Now therefore, we always allocate a standard size, either 64K or 2MB when extending the HSA.

The test appends a few large strings and calls checkConsistency() to chekc the arenas.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5921

Reviewed By: xiaoxmeng

Differential Revision: D47905156

Pulled By: oerling

fbshipit-source-id: da08ac2a19ab1e0b49915ed35161c5982b41c8a2