Fix macOS CircleCI builds (#5937)
Summary:
Currently mac os circle Ci builds are failing as they do not have ccache
installed. The installation script is not executed if a previous cache
with installed dependencies exists. Ideally ccache should be there in
the cache but its not, so we always run the setup script incase
something fails to be included in the cache.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5937

Reviewed By: kagamiori

Differential Revision: D47959519

Pulled By: bikramSingh91

fbshipit-source-id: 3e42c0cb4e853b0bc15464434e86c6d25895e1e0