Fix segment fault while creating flat vectors (#4960)
Summary:
We meet this problem while running with gluten, and we found that the array was out of bounds while getting the last offset. The length of the offsets is `length+1`, thus the last index should be `length` instead of `length+1`. As for the unit tests, they are running normally because it use offset to read data and won't check whether has too many data in stringViewBuffers.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4960

Reviewed By: xiaoxmeng

Differential Revision: D48402244

Pulled By: Yuhta

fbshipit-source-id: 0207ab9739f62371b26c3f4fd48da903cc15401d