Fix race condition in SharedArbitratorTest.reclaimFromJoinBuilder* tests reading/writing to QueryConfig (#6188)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6188

These tests use AssertQueryBuilder to update the QueryConfig in the QueryCtx object, this runs in one thread.  Unfortunately the same QueryCtx may be used to run another query in a second thread.  This can lead to the QueryConfig being read and written at the same time leading to undefined behavior.

The fix is to just update the QueryConfig before we start running any of the threads.

Reviewed By: xiaoxmeng

Differential Revision: D48538101

fbshipit-source-id: ebce778366899cd54d619a84949bc383a3bf30ba