Fix the memory budget check in dwrf writer flush control (#6227)
Summary:
Currently dwrf relies on the capacity limit in dwrf writer pool to detect if it
needs to flush or not. The capacity limit in dwrf writer pool is actually the
capacity from the associated query root pool as velox doesn't do quota
enforcement at operator or plan node level.

But we should not check on the capacity, as with memory arbitrator,
the capacity is the memory pool's current capacity and is different
than the max capacity. The former can grow up to max capacity.
The dwrf writer flush also depends on the control logic in flush policy.
Without this fix, we might see flush becoming more often with
memory arbitration enabled as the pool's initial capacity could be small.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6227

Reviewed By: HuamengJiang, kewang1024

Differential Revision: D48614527

Pulled By: xiaoxmeng

fbshipit-source-id: 56f77ad4dea59109da82c660ea3c80d2bba9d5a5