compare function return types when sorting signatures (#6315)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6315

Today, when we setup the ExpressionFuzzer we sort the list of signatures.  Currently we do this by
comparing the names, the number of arguments, and the argument types.

Unfortunately, cast throws and wrench in the works, and while this would normally be sufficient, we can have
two cast calls where all of these are equal, but the signatures have different return types.  (Oddly there's a
comment above in the code saying that the return types will be compared even though they aren't).

I tracked down an issue reproducing a failure using a given seed to this.

Reviewed By: bikramSingh91, mbasmanova

Differential Revision: D48762001

fbshipit-source-id: 9d8431566f27d936d24565a0efe3beb9a4f03395