Fix StringBuffer re-use in PrestoSerializer (#6337)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6337

When generating a string flatvector from a presto page, the presto
serializer looks for an existing stringBuffer in the FlatVector to
re-use for appending values. It does that by finding the smallest
available buffer that has enough 'capacity' to accommodate all the
values in the incoming presto page. However, currently it does not
update the 'size' the of buffer which it ends up re-using. This
results in all the newly appended stringView's pointing outside the
'size' but within the 'capacity' of the buffer. This patch fixes
this discrepancy.

Reviewed By: kgpai

Differential Revision: D48800168

fbshipit-source-id: d35ac648950ce4df4f2430c73330385253f6936b