SelectiveFlatMapColumnReader can set arbitrary maps to null (#6351)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6351

There's a bug in SelectiveFlatMapColumnReader where if it encounters any maps where none of the keys match the filter, it will
arbitrarily set maps in the batch to null.

When SelectiveFlatMapColumnReader encounters a map that would be empty because of the key filter, setNulls with a null buffer
containing uninitialized values.  I think the intention was to set the current map to null in the results.

Since the null values aren't initialized and since we never unset nulls, maps throughout the batch are arbitrarily set to null (notably
non-null maps that we've already read).  Since we're using rawResultNulls_ to tell if the flattened struct is null in the file, this
overwrites that and we now read some maps later in the batch as null that should aren't and some maps that aren't null as null.
Moreover, setting empty maps to null, which I think was the intention, is incorrect anyway as they're empty, not null.

The fix is to move this code to where we check for nulls, and to check for nulls using the correct fields.

Reviewed By: Yuhta

Differential Revision: D48835969

fbshipit-source-id: 543e7c7bb155bf622fb8f36f4a70e79bf85729b2