Fix map_from_entries top level row nulls handling. (#6391)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6391

DecodedVector->wrap() does not propagate base vector nulls, fixing it is not trivial and  introduces several issues
will take longer time (https://github.com/facebookincubator/velox/pull/6376). This is a hot fix for the function to
mitigate the issue from production and from fuzzer runs. Also general there is a comment that says that
DecodedVector->wrap() should not be used outside the PeelEncoding context.

This fix the issues in fuzzer issues #6374 and #6255. The problem there is that we have
row top level nulls that need to be set as nulls in the wrappedKeys and wrappedValues but they did not use to before
this diff.

Reviewed By: bikramSingh91

Differential Revision: D48910201

fbshipit-source-id: 7d50767f66eb29d673c373afd693fd33ee0937bd