Add ability to generate nested lazy children for RowVectors (#6109)
Summary:
This adds the ability to add lazy children nested withing row vectors.
It builds on top of the existing lazy vector generation mechanism in
fuzzer that ensures deterministic repro on failures by keeping track
of the input columns that are to be wrapped. If the column to be
wrapped is a row vector, it checks whether it can recursively
traverse its children to wrap the innermost child.

Found a bug https://github.com/facebookincubator/velox/issues/6106 with this change and it therefore is blocked
on it to be merged.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6109

Reviewed By: kgpai, kagamiori

Differential Revision: D48401127

Pulled By: bikramSingh91

fbshipit-source-id: 799a9b4d1a6878c4d1ee52182fa519fdfd9da7a1