Handle inner-most lazy layer of input vector in CastExpr (#6383)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6383

PeeledEncoding doesn't peel off inner-most lazy layer, so CastExpr can get a lazy vector after peeling. But the current code assumes that the vector after peeling is always a simple vector and this caused crashes. This diff fixes this bug by loading the vector after peeling if it is lazy.

Reviewed By: bikramSingh91

Differential Revision: D48890110

fbshipit-source-id: 7d56f51337f21b4c8db48870f24b446e8c82e9d7