Add S3FileSystem Write support (#6021)
Summary:
S3WriteFile uses the Apache Arrow implementation as a reference.
AWS C++ SDK allows streaming writes via the MultiPart upload API.
Multipart upload allows you to upload a single object as a set of parts.
Each part is a contiguous portion of the object's data.
While AWS and Minio support different sizes for each
part (only requiring a minimum of 5MB), Certain object stores require that every
part be exactly equal (except for the last part). We set this to 10 MiB, so
that in combination with the maximum number of parts of 10,000, this gives a
file limit of 100k MiB (or about 98 GiB).
You can upload these object parts independently and in any order.
After all parts of your object are uploaded, Amazon S3 assembles these parts
and creates the object.
S3WriteFile is not thread-safe.
UploadPart is currently synchronous during append. Flush is no-op as append
handles all the uploads.

Resolves: https://github.com/facebookincubator/velox/issues/4805

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6021

Reviewed By: kgpai

Differential Revision: D49324662

Pulled By: pedroerp

fbshipit-source-id: f26479058f576a63f7d4fe4527b57bd0aa87ab30