Track number of rows processed in Window::getOutput (#6584)
Summary:
Streaming window needs the number of rows processed in the getOutput() call for
the result vector. Not all received rows might be assigned to a partition yet.
In fact, the last partition is always open until noMoreInput is received. So
getOutput() can only process all closed partitions which is less than the
number of input rows.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6584

Reviewed By: amitkdutta

Differential Revision: D49330430

Pulled By: mbasmanova

fbshipit-source-id: 7b54b1c95fdeb496f734916d7f26cb14c095f71d