Add ability to serialize input vector and sql if expression evaluation crashes (#6402)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6402

This adds an experimental flag
'experimental_velox_save_input_on_fatal_signal' that when set to
true, serializes the input vector data and all the SQL expressions
in the ExprSet that is currently executing whenever a fatal signal
is encountered. Enabling this flag makes the signal handler async
signal unsafe, so it should only be used for debugging purposes.

Reviewed By: kgpai

Differential Revision: D48891649

fbshipit-source-id: 47722d726c76a8602cf436c1840d2a0d720e2c35