Add field separators as parameter in Subfield (#6014)
Summary:
In Velox tokenizer, some characters, eg. DOT, are treated as separation
characters by default. But in Spark, DOT is treated as a regular character when
it exists in column name.This PR added field separators as a parameter of
Subfield to allow configurable behaviors. Separation happens when the input
char is a separator and can be found in the path. Otherwise, the char is not
used to tokenize.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6014

Reviewed By: xiaoxmeng

Differential Revision: D49331238

Pulled By: Yuhta

fbshipit-source-id: 445e343cfc264e8c0720eb58525727a999e596d0