Fix trim with leaf invalid unicodes. (#6684)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6684

Recent change to isUnicodeWhiteSpace makes it segfault if a negative number is passed to it.
function could call isUnicodeWhiteSpace with -1 when the input is invalid unicode.
I also added explicit check in trim not to call it with invalid, but it's good to have the checks in both places.

Reviewed By: mbasmanova

Differential Revision: D49510430

fbshipit-source-id: 2b7f6ba6de096edc6bcf69801b3d2d4dcf05c653