Fix empty sort buffer spill check failure (#6840)
Summary:
With memory arbitration, we might trigger spilling on an order by
operator even if it has no spillable data, and this cause a check failure
discovered in Meta internal testing.
This PR fixes this issue by avoiding spilling if order by operator or the
sort buffer is empty.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6840

Reviewed By: amitkdutta

Differential Revision: D49804183

Pulled By: xiaoxmeng

fbshipit-source-id: d5797c52ca24c40da71fc0583c5bb91b5e284c93