Add spill memory pool check for memory reservation under arbitration (#6860)
Summary:
Add check in memory reservation if the request memory pool is the spilling
memory pool or not under memory arbitration. We shall only allow memory
reservation from a spilling memory pool during memory arbitration. We add
MemoryArbitrationContext for this which is set o a per-thread local object, and
check this on memory pool reservation request.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6860

Reviewed By: mbasmanova

Differential Revision: D49853858

Pulled By: xiaoxmeng

fbshipit-source-id: 6d21191774b50aa4c207b982950eb0e26aa78253