Handle exception from remote split (#6752)
Summary:
During task termination, if there are remote splits remaining, they
will need to be added to the task before the task ends. However, if
`addRemoteSplit` throw an exception. The task termination process will
be interrupted and the cleanup afterward will not be able to execute.
This is problematic when there is unfinished blocked Join operator in
the task, as it relies on `bridge->cancel()` at the end of the task
termination process to clean its unfilfulled promise. If exception
happens, `bridge->cancel()` will not be called and the driver cannot be
cleaned up, resulting in zombie task hanging.

This change handles possible exception from `addRemoteSplit` so that the
cleanup logic will be executed properly to prevent zombie tasks.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6752

Reviewed By: xiaoxmeng

Differential Revision: D49854355

Pulled By: zacw7

fbshipit-source-id: 2a8f07c39d8c626190b246e064c9a70f4ca54990