Fix flaky test for zombie task (#6892)
Summary:
When check `ASSERT_EQ(task.use_count(), 1)` is enforced, it is possible
that `waitForTaskFailure` has not finished, holding the task reference
and causing the test failure. Instead of checking task.use_count(), we
can rely on `QueryAssertions#waitForAllTasksToBeDeleted` to check if
the task is hanging at the end of the test.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6892

Reviewed By: xiaoxmeng

Differential Revision: D49915580

Pulled By: zacw7

fbshipit-source-id: d027877012dac7178c728548158e749bcda8514a