compareMaterializedRows should compare unique row counts (#6919)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6919

compareMaterializedRows in QueryAssertions currently checks that the number of rows in Velox and DuckDb
match, and that every row output by one appears in the other's results.

This is not strict enough, as query results may have duplicate rows.  compareMaterializedRows will miss the
case where the total number of rows is the same but the number of each distinct row is different.

The fix is to check that the counts match, not just that the row is present.

Reviewed By: kagamiori

Differential Revision: D49895573

fbshipit-source-id: 2f058f8de9f58b6961ded329fe1c79cf32f68127