Fix: Avoid calling resize in addNulls for vectors of all encodings not only flat. (#6958)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6958

bikramSingh91 is working on a refactor to optimize AddNulls, in the meanwhile this diff
fix an exisiting bug, the previous fix ( https://github.com/facebookincubator/velox/pull/6856)
only apply to flat row, this make it work for all encodings.

To explain the error
- say input is Row of size 3 wrapping in dict.
- we call ensure writable then resize we get
row: size (5)
child size: 3.
- then we add nulls only to row 4, because 3 is not selected.
in that case the result vector is not valid and accessing row number (3) will
be invalid.

Fix https://github.com/facebookincubator/velox/issues/6935

Reviewed By: kagamiori

Differential Revision: D50062849

fbshipit-source-id: e4d84c04eae8448a7fc82e489f6d041a23b5fac9