Optimize WindowOperator for pre-sorted input (#5437)
Summary:
Spark planner inserts a Sort operator before a Window operator to sort the data.
Hence, there is no need to sort the data again in the window operator.

Add an option to WindowOperator to skip sorting if inputs are already sorted.
This allow WindowOperator to process data in streaming manner without
accumulating all input in memory.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5437

Reviewed By: laithsakka

Differential Revision: D50198545

Pulled By: mbasmanova

fbshipit-source-id: 05ed9ea26691c00310730ceea0153405291023ab