Fix TEXTFILE reading when filter on partition key but not project out (#7057)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/7057

TEXTFILE table is pushing down partition key filters (not project out) down to the worker (DWRF table is not doing that in contrary).  These columns are not in data columns so it will give us an error when we try to get their type and constructing subfields pruning for them.  To handle this, we need to exclude them from subfields pruning construction.

Reviewed By: oerling

Differential Revision: D50284303

fbshipit-source-id: f51b3d24750e9c41a6c9b954d0d21217a2fbb338