PeeledEncoding to accept DecodedVector as opposed to LocalDecodedVector (#7098)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/7098

LocalDecodedVector is only ever used as DecodedVector in PeeledEncoding methods, so the signature should really not restrict to LocalDecodedVectors. This also allows other use cases to use encoding peeling to help vector transforms.

Due to the requirement of calling get() for the conversion, I'm adding an additional api for smaller blast radius.

Reviewed By: Yuhta

Differential Revision: D50376955

fbshipit-source-id: 65ac2beddb7829798f5aa71a6528aa6ecf08276e