Fix MultiFragmentTest.maxBytes unit test (#7104)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/7104

Currently the test relies on PartitionedOuput operator to enforce a
1MB limit on the output batches however the operator also has a
limit on the number of rows which varies randomly between 70% to
120% of 10000. This limit was not correctly enforced and was fixed
in #7059 so after it was merged this randomness resulted in this unit
test being flaky as it utilized a vector of size 10000 and expected
the output to be predictable.
The fix here is to reduce the row count of the input vector while
ensuring its memory usage stays the same in order to hit the memory
limit that the test is checking for.

Reviewed By: xiaoxmeng

Differential Revision: D50388151

fbshipit-source-id: 77a08c1f95c7dabb2995af3fc89f5774a59e4dc3