Make RowType's contructor throw VeloxRuntimeError. (#7156)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/7156

Make RowType's contructor throw VeloxRuntimeError
instead of VeloxUserError on "Mismatch names/types sizes".
We are seeing this error in batch workload and it is not
readily reproducible, which makes it a Runtime error.
We also want to see more details on the names and types.
Finally, we want to see the stack and maybe save
Vector + Expression.

Reviewed By: gggrace14

Differential Revision: D50472707

fbshipit-source-id: bb0b452e4c1c82f2f7bba359cd726c616a60a951