Fix setNullField in DWRF struct reader to handle missing subfield in file and newly created result (#7129)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/7129

Currently if a leaf level nested subfield is missing in file, and we
cannot reuse the result, the `VectorPtr` is null at that child and we try to
access type and pool information out of it.  Fix this by getting the type and
pool from the parent vector instead.

Reviewed By: oerling

Differential Revision: D50422967

fbshipit-source-id: ba7fd08f4399bacd5ea45f4a30b55beb6e810ce3