Adopt spill-friendly order of input columns in TopNRowNumber operator (#7144)
Summary:
TopNRowNumber operator maintains a hash table of partitions and a RowContainer
of input rows (up to N rows per partition).

To add spilling support, it is convenient to store columns in the order of
partition keys, followed by sorting keys, followed by the rest of the input
columns.

This column order allows to sort and spill accumulated inputs, then merge-sort the
data back and produce the output.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7144

Reviewed By: xiaoxmeng

Differential Revision: D50456953

Pulled By: mbasmanova

fbshipit-source-id: 6e6f1c7a46d0ad1ef75f9b0efe7a47e3c619f973