Rewrite and optimize EvalCtx::addNulls method (#7190)
Summary:
This change rewrites the EvalCtx::addNulls() implementation to ensure
all cases are covered for generating a valid output vector.
Additionally adds an optimization where it wraps complex vectors
in a dictionary instead of copying them for cases where they are
not mutable or need to be resize.

Additionaly, modified tests that relied on the output to be a
non-encoded (dictionary or constant) complex vector.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7190

Test Plan: Verified that all tests pass

Reviewed By: Yuhta

Differential Revision: D50563551

Pulled By: bikramSingh91

fbshipit-source-id: fa4a8e813e95c06764cd419c96574cd2deb0dad8