Fix metadata computation for shared lambda expressions (#7265)
Summary:
Currently the distinct fields for lambda expression are updated only
during construction. However, this can be an issue if we want to
clear and recompute it which happens when the expression is marked
as a shared expression. This change fixes it by ensuring recomputing
correctly updates the distinct fields every time.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7265

Test Plan: Added unit test

Reviewed By: mbasmanova

Differential Revision: D50714166

Pulled By: bikramSingh91

fbshipit-source-id: 4db510983c2432296038972686d67e3edffa0d82