Avoid reading index stream before loading (#7236)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/7236

In `SelectiveStringDictionaryColumnReader`, we read the row group index
stream in constructor by mistake.  This caused failed reading when
`BufferedInput` is used (`CachedBufferedInput` would work but performance is
suboptimal).  The row group index is already ensured to be loaded in
`SelectiveStringDictionaryColumnReader::ensureInitialized` which is the right
time for it to happen (after loading).

Reviewed By: oerling

Differential Revision: D50658787

fbshipit-source-id: 5ab6a5252f5dcb48c921e00d49f56f50f7797bb7