Add id to SpillMergeStream (#7315)
Summary:
The id to SpillMergeStream to help order the spilled file from a given
spill partition. The spilled file id starts from zero of the first spilled file.
The in-memory spill stream has a special max uint32_t value. This
helps identify the first spilled file used in un-spilling merge path for
special handling such as distinct aggregation used to tell if a output
group is distinct or not.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7315

Reviewed By: tanjialiang

Differential Revision: D50802089

Pulled By: xiaoxmeng

fbshipit-source-id: f9e111ca629abd11f3e0d713e8a0ec532c9b9176