Expressions: add explicit operator== overloads for the concrete type
Summary:
The current operator== works with generic types, but not when lhs and rhs have the same type. In that case, C++ complains about ambiguous operator== overload.

This is uncovered when switching to googletest v1.14, which starts to use std::equal_to to compare two variables of the same type.

To fix, add explicit operator overloads for the concrete types and have the generic operator== instances call these overloads for better code reuse.

Reviewed By: kalman5

Differential Revision: D50836472

fbshipit-source-id: 9a2511bac9121ac901aa3d9223ee4bd0b5e64c26