Revert D50803466: Multisect successfully blamed "D50803466: [velox][PR] Faster StringView hash" for test or build failures
Summary:
This diff is reverting D50803466
D50803466: [velox][PR] Faster StringView hash by oerling has been identified to be causing the following test or build failures:

Tests affected:
- [koski/py/tests:hash_udf_test - test_binary (koski.py.tests.hash_udf_test.HashUdfTest)](https://www.internalfb.com/intern/test/844425003260387/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3440725
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: kevinwilfong

Differential Revision: D50858317

fbshipit-source-id: bf9670f6f402dfcbebbe34cb40641fcdfb6d2f4b