Add row vector unsafe resize api. (#7198)
Summary:
This PR adds the row vector unsafe resize api. This Api is only intended for backwards compatibility in certain sensitive places (for e.g I/O Reader/Writers). The next PR will add recursive resize to all children of a RowVector.
As a part of this change we also make changes to certain sections of the code which rely on the old behavior and cannot be immediately migrated away.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7198

Reviewed By: laithsakka, bikramSingh91

Differential Revision: D50577808

Pulled By: kgpai

fbshipit-source-id: 6fc86f84902a77d953aa3c0b13371fc67f945d08