Optimize element_at for maps with complex type keys (#7365)
Summary:
Implementation of element_at(map, search) uses linear search to find matching
entry. This is slow when maps are large (100s of entries). This optimization
sorts map keys and uses binary search producing 4x speedup on a real-world
workload. The optimization applies only if there is a single map (constant or
dictionary encoded) since in this case the cost of sorting the map amortizes
across many searches.

There is a standard query optimization technique that generates element_at
over a single map.

A simple join with a small lookup table u:

```
SELECT * FROM t, u WHERE t.id = u.id
```

Can be written as

```
SELECT
   element_at(
      (SELECT map_agg(u.id) FROM u), t.id
   )
FROM t
```

This ensures that records from u will be collected into a single map and
broadcasted to all workers to be "joined" with t. The original join query may
run as partitioned join (if optimizer is not smart enough) and be very slow.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7365

Reviewed By: Yuhta

Differential Revision: D50882410

Pulled By: mbasmanova

fbshipit-source-id: 74d00fae7417f128656e3bdc1b5e2c91060e96a7