Improve BenchmarkParquetReader (#6275)
Summary:
This PR is to fix https://github.com/facebookincubator/velox/issues/6247.

So far the Benchmark test has 4 types, this PR added more types to the BenchmarkParquetReader test:

1. ShortDecimalType
2. LongDecimalType
3. VARCHAR

We should test them with different filter rates and null rates just like the previous tests in BenchmarkParquetReader.

Also, I noticed we have never done the filter test for the type HUGEINT (int128_t) before, which is the actual type of LongDecimalType. So this PR also implemented the code to generate the filter of the type HUGEINT (int128_t) and modify the code to generate the data of the type HUGEINT (int128_t) correctly.

While developing the code, we found there is another problem/enhancement we need to do, so I created an issue for it https://github.com/facebookincubator/velox/issues/6248 This could be the next step.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6275

Reviewed By: pedroerp

Differential Revision: D50934566

Pulled By: Yuhta

fbshipit-source-id: 12f73995b091448dbd9a69cddd3d8eae0222d3f2