Add spilling support for distinct aggregation (#7305)
Summary:
Add spilling support for distinct aggregation:
(1) when spilling has triggered, the hash aggregation operator stops producing the
distinct output rows;
(2) the spilling write out all the rows in row container to disk;
(3) during the output stage, we leverage the stream id from the spill file to tell if the
spilled row is a new distinct or not. If a spilled row contains in the first merge stream,
then it is not a new distinct, otherwise it is a new distinct.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7305

Test Plan:
Failed OOM query: 20231103_054044_00000_qqqst
Succeeded one with this protection: 20231103_060612_00000_8vuy5

Reviewed By: mbasmanova, oerling

Differential Revision: D50781129

Pulled By: xiaoxmeng

fbshipit-source-id: 90559e3e09838aea8ca8116a2720e736ddfcb1c6