Fix CPU regression (I doubt this is the cause though) (#7430)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/7430

A CPU regression in `facebook::velox::dwrf::DwrfRowReader::readNext` was flagged (as regressing 30%) with the introduction of parallel decoding. Even when we're not using parallel decoding. It looks quite impossible for me by the way, but maybe a mess with the branch prediction could cause it.

The only thing that I can think of is that we check for the existence of the executor inside a loop. I assumed that the branch prediction would take care of it, since it will be always true or false. But maybe it isn't. Let's move it outside the loop and see what happens.

Reviewed By: Magoja

Differential Revision: D51037397

fbshipit-source-id: 9b567b88c8d7d861a421785587ec9f056cb43190