Fix empty aggregation spill check failure (#7466)
Summary:
The spilling can be triggered during aggregation output processing stage.
If it happens that all the output have been produced and before we finish the
operator, the output spilling will be triggered but nothing will be spilled. This
triggers a check failure in un-spill path which expect spill data if the spiller has
been triggered.
This PR remove this check and adds unit test to reproduce this. Note it is not
easy to detect if a hash aggregation operator has produced all the output or not
from row container iterator.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7466

Reviewed By: tanjialiang, mbasmanova

Differential Revision: D51099553

Pulled By: xiaoxmeng

fbshipit-source-id: e491e2af25b27a78f4f39cc11eeef08a50d13aed