Skip spilling if all the partitions have been spilled (#7476)
Summary:
If memory arbitration has reclaimed memory from the hash build operator which
request memory reservation, and there is still lack of memory, then the hash build
operator will try spill. However, this is not required if the memory arbitration
has enabled. This can lead to a check failure in HashBuild::runSpill which expects
not all the partitions have been spilled.
This PR adds a check if all the partitions have been spilled when memory reservation
fails for a hash build operator, and skip spilling if all the partitions have been spilled.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7476

Reviewed By: mbasmanova

Differential Revision: D51137514

Pulled By: xiaoxmeng

fbshipit-source-id: 23f2aceedfadea52c054dd91bccb1754f2bc5b67