Add reserve and release counter in SharedArbitrator::Stats (#7530)
Summary:
MemoryManager calls `SharedArbitrator::reserveMemory` each time
a root memory pool is created by its API `addRootPool`, and calls
`SharedArbitrator::releaseMemory` through a callback method in the
destructor of the `MemoryPoolImpl` when destructing a memory pool.

This PR adds two counters and for `SharedArbitrator::reserveMemory`
and `SharedArbitrator::releaseMemory` to count their call counts,
which should be the same, at the end of the `MemoryManager` life cycle.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7530

Reviewed By: tanjialiang

Differential Revision: D51242751

Pulled By: xiaoxmeng

fbshipit-source-id: 5de0274b42b0264bedf0b19a5e81fc7fb1c2a85a