Add zerocopy append and bit bulk append to ByteStream (#7731)
Summary:
Adds fast paths for fixed size append of T and bits. This is needed for zerocopy append of arrays of different types to a ByteStream. Otherwise a temp copy would always have to be allocated and filled and then appended. Now a temp copy is made only if the contiguous space in the ByteStream does not cover the append.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7731

Reviewed By: Yuhta

Differential Revision: D51577228

Pulled By: oerling

fbshipit-source-id: 5d588f8373e5029f243f8d5392fc3c2e348becb1