Fix memory related flaky tests (#7815)
Summary:
Two flaky tests are caused by recent optimization on memory reservations for
empty hash table and the test condition changed:
HashJoinTest.reclaimDuringReserve
SharedArbitrationTest.reclaimFromJoinBuildWaitForTableBuild
Fixed the test condition plus some code cleanup

One is tsan flakiness but it actually can cause problem in prod as spiller is not
thread safe when we check if we can spill, spiller might be set in hash build
operator but when we actually spill from the operator, there is possibility (in theory)
that the spiller has been cleared. So it might be safe to check if spill has been
configured and actually check if spiller object is set or not when we actually do
memory reclaim.

And a bug fix in AssertQueryBuilder

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7815

Test Plan:
https://www.internalfb.com/intern/testinfra/testrun/4222124861182357
https://www.internalfb.com/intern/testinfra/testrun/15199648754974678

Reviewed By: mbasmanova

Differential Revision: D51729742

Pulled By: xiaoxmeng

fbshipit-source-id: 15a7910ac41c381b885edc75de16ed8e0735d471