Make ByteStream::size() consistent across use cases (#7520)
Summary:
ByteStream::size is expected to have the semantics of the length of std::stringstream. It is the offset of the first unwritten byte from the start. This does not move backward when seeking to mid stream and writing, for example.  been written.

Adds a test to verify proper function when randomly overwriting a ByteStream backed by HashStringAllocator multipart entries.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7520

Reviewed By: xiaoxmeng

Differential Revision: D51270817

Pulled By: oerling

fbshipit-source-id: 7b772852663b933b600c1d52b400a37ea9429edc