Prevent infinite looping from hash table probe operations (#8012)
Summary:
There are infinite loops in hash table probe operations that
break until find the search entry or find out an empty slot.
This could result in infinite loop when there are some bugs in
the relevant code path. To prevent this, we shall break out of
the loop if we have searched through all the buckets in the
table

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8012

Reviewed By: mbasmanova

Differential Revision: D52110273

Pulled By: xiaoxmeng

fbshipit-source-id: 225d688a0d0250e6f49601ec0985dd5c12073f2d