Fix variance edge case (#8028)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/8028

When the current accumulator is empty, and then we are adding another
accumulator with `m2` is infinity and `mean` with a very large value that when
squared would become infinity, the new `m2` calculation is indeterminate ($` \infty \times 0`$) and becomes `NaN`.  The correct new `m2` should be infinity.  We should just
take the values of second accumulator directly in this case.

Also update the merge function of covariance just to be safe.

Reviewed By: tanjialiang

Differential Revision: D52132799

fbshipit-source-id: 191e099ed0bb1dcf7f812e787772dedac122bcd2