Replace the overuse of StringView with std::string_view in Re2Function (#8047)
Summary:
Currently there are some overuse of StringView in Re2Functions.h/cpp. It's not good because StringView is not a replacement of std::string_view, its copy is expensive than std::string_view, its data() is not safe to point to for copied short StringView.

In this PR we replace StringView with std::string_view if it is a field of struct or a function param which is not feed by XxxVector directly.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8047

Reviewed By: Yuhta

Differential Revision: D52167089

Pulled By: mbasmanova

fbshipit-source-id: 63e2fdf85291a7ab67c9cde39f0ae7c5445a2d10