Refactor spiller constructors (#8184)
Summary:
The spiller public constructors contain many members
of the `spillConfig`, it would be nice to pass the `spillConfig`
and use its members as the spiller's private constructor parameters.
Remove the memory pool parameters from the spiller constructors
as the memory pool used by the spillers is the same one created by
the singleton method `memory::spillMemoryPool()`.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8184

Reviewed By: bikramSingh91

Differential Revision: D52432446

Pulled By: xiaoxmeng

fbshipit-source-id: bd6f0c9c6ad953979ea3084e107ef14b18fb9f47