Fix estimateRowSize with wrapper (#8199)
Summary:
Fix estimateRowSize with wrapper

When inputs of PartitionedOutput are wrapped, e.g. because of a
previous filter, the per row size estimation was 0, leading to
excessive buffering on the producer. The flush was then only triggered
by row number thresholds, which would be bad for wide rows.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8199

Reviewed By: xiaoxmeng

Differential Revision: D52450236

Pulled By: oerling

fbshipit-source-id: 926593ad416092eb1e5a6bc830910262a35903ea