Guard against AppendWindow destructor throw (#8203)
Summary:
AppendWindow is a scoped guard over a fixed size append to a ByteStream. If the ByteStream does not have space, then a scratch piece of memory is used to provide contiguous space and is then appended to the ByteStream on exit. This last append may however fail. This may throw from stack unwinding and kill the process.

To prevent this, we ensure that all the memory is available at AppendWindow construction time. The destructor cannot throw because of failure to extend the stream. We add a LOG(FATAL) to mark the unreachable throw from destructor.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8203

Reviewed By: xiaoxmeng

Differential Revision: D52458620

Pulled By: oerling

fbshipit-source-id: 7daf4be3277af4825b5d948e8d5e3998d3dbe592