Remove tests of debug time checks (#8261)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/8261

Its is not possible to cast to two different primitive (one of them is def of a wrong type). In debug time this would be caught with a debug time check, but that check is
 not needed in opt time, since castTo assumes the users know the casting is valid
(unlike tryCastTo).

What used to happen is that another check used to be triggered that (the one checking that we are always casting to the same type) but that check is not needed for primitives and a previous diff added a path that does not perform that check,.

Reviewed By: kgpai

Differential Revision: D52544148

fbshipit-source-id: 83499ba6e8855ec517842991053d0b1e8640a6ea