Add support for allocations that exceed the PageRun limit (#8298)
Summary:
This change enables an allocation to exceed the PageRun limit. We do this by splitting an allocation
across multiple PageRuns.
Non-contiguous allocations in the MallocAllocator use malloc to allocate and store the address
to later verify the allocation. Since an allocation can now span across multiple PageRuns, we need to
store the MachinePageCount for each allocation to identify allocation boundaries across the PageRuns.
Fix stats for `recordFree` in MallocAllocator nonContiguousFree.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8298

Reviewed By: oerling

Differential Revision: D52611261

Pulled By: xiaoxmeng

fbshipit-source-id: 9d09d4c4aaaed3228206bf367591e6a0fbf34dbe