Preserve recursive encodings while serializing PrestoPage (#8249)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/8249

In the current PrestoPage serialization code, only the first level of
encodings (dictionaries and constants) are preserved, since at times
VectoStreams are created without the actual data - so that the
encoding is unknown.

When data is serialized using `serializeEncoded()` and hence the vector data
is available, we can recursively capture the encoding and create the appropriate
streams to serialize dictionaries and constants, thus preventing us from
flattening them.

Reviewed By: kparichay, mbasmanova

Differential Revision: D52524708

fbshipit-source-id: 10b77d4ff89ed54323f8c4c0ebda9fd671ee765c