Fix min/max(x, n) (#8311)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/8311

Velox has an optimization for Window operation with incremental aggregation when there are peer
rows with the same frame. In this situation, the aggregation result is only computed at the first row
of the peer and the rest rows simply copy this result. This optimization assumes that results of
incremental aggregation in Window operation on peer rows should be the same. However, min/max(x, n)
in Velox breaks this assumption because their extractValues() method causes the accumulator to
be cleared, making the peer rows after the first row expect a different result. This diff fixes min/max(x, n)
to make the extraction method not clear the accumulator. The behavior after the fix also align with
Presto's.

This diff also adds a method testIncrementalAggregation in testAggregations to check that extractValues()
doesn't change accumulator afterwards for all aggregation functions. After this fix, only min_by/max_by(x, y, n)
doesn't pass testIncrementalAggregation.

This diff fixes https://github.com/facebookincubator/velox/issues/8138.

Reviewed By: mbasmanova

Differential Revision: D52638334

fbshipit-source-id: 7509b713a34ffa03a35dbeb2751e714b2203e302